Stripping CommandLineParser out of application.
