+++ meson.build
@@ -267,7 +267,7 @@ lib_src += [
   'src/library/tools/activatable.cc',
   'src/library/tools/application/application.cc',
   'src/library/tools/application/init.cc',
-  'src/library/tools/application/cmdline.cc',
+  'src/library/tools/commandlineparser.cc',
   'src/library/tools/application/run.cc',
   'src/library/tools/application/setup.cc',
   'src/library/tools/arguments.cc',
@@ -801,6 +801,7 @@ install_headers(
   'src/include/udjat/tools/timestamp.h',
   'src/include/udjat/tools/value.h',
   'src/include/udjat/tools/application.h',
+  'src/include/udjat/tools/commandlineparser.h',
   'src/include/udjat/tools/exception.h',
   'src/include/udjat/tools/factory.h',
   'src/include/udjat/tools/file.h',
